fix(cli): Remove circular import by extracting function
