+++ .changeset/khaki-nails-know.md
@@ -1,0 +1,5 @@
+---
+'@backstage/cli': patch
+---
+
+Removed circular import
