+++ packages/cli/src/commands/versions/migrate.ts
@@ -19,7 +19,7 @@ import { resolve as resolvePath, join as joinPath } from 'path';
 import { OptionValues } from 'commander';
 import { readJson, writeJson } from 'fs-extra';
 import { minimatch } from 'minimatch';
-import { runYarnInstall } from './bump';
+import { runYarnInstall } from './utils';
 import replace from 'replace-in-file';
 
 declare module 'replace-in-file' {
